cambiar generated value a sequence, y modificar el incremento a 1.

+++ src/main/java/org/cslt/movies_app/controller/MovieController.java
@@ -20,7 +20,7 @@ public class MovieController {
     }
 
     @GetMapping("/{id}")
-    public Movie getMovie(@PathVariable Long id) {
+    public Movie getMovie(@PathVariable int id) {
         return movieService.getMovieById(id);
     }
