+++ src/main/java/org/cslt/movies_app/services/MovieService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface MovieService {
 
     List<Movie> getAllMovies();
-    Movie getMovieById(Long id);
+    Movie getMovieById(int id);
 
 }
